# Add hook support