--- a/event.go
+++ b/event.go
@@ -23,11 +23,12 @@ type hookRunner func(e *Event, level Level, msg string)
 // Event represents a log event. It is instanced by one of the level method of
 // Logger and finalized by the Msg or Msgf method.
 type Event struct {
-	buf   []byte
-	w     LevelWriter
-	level Level
-	done  func(msg string)
-	hr    []hookRunner
+	buf      []byte
+	w        LevelWriter
+	level    Level
+	done     func(msg string)
+	h        []Hook
+	hasLevel bool
 }
 
 // LogObjectMarshaler provides a strongly-typed and encoding-agnostic interface
@@ -48,7 +49,7 @@ func newEvent(w LevelWriter, level Level, enabled bool) *Event {
 	}
 	e := eventPool.Get().(*Event)
 	e.buf = e.buf[:1]
-	e.hr = e.hr[:0]
+	e.h = e.h[:0]
 	e.buf[0] = '{'
 	e.w = w
 	e.level = level
@@ -79,11 +80,11 @@ func (e *Event) Msg(msg string) {
 	if e == nil {
 		return
 	}
-	if len(e.hr) > 0 {
-		e.hr[0](e, e.level, msg)
-		if len(e.hr) > 1 {
-			for _, hook := range e.hr[1:] {
-				hook(e, e.level, msg)
+	if len(e.h) > 0 {
+		e.h[0].Run(e, e.level, e.hasLevel, msg)
+		if len(e.h) > 1 {
+			for _, hook := range e.h[1:] {
+				hook.Run(e, e.level, e.hasLevel, msg)
 			}
 		}
 	}
@@ -107,11 +108,11 @@ func (e *Event) Msgf(format string, v ...interface{}) {
 		return
 	}
 	msg := fmt.Sprintf(format, v...)
-	if len(e.hr) > 0 {
-		e.hr[0](e, e.level, msg)
-		if len(e.hr) > 1 {
-			for _, hook := range e.hr[1:] {
-				hook(e, e.level, msg)
+	if len(e.h) > 0 {
+		e.h[0].Run(e, e.level, e.hasLevel, msg)
+		if len(e.h) > 1 {
+			for _, hook := range e.h[1:] {
+				hook.Run(e, e.level, e.hasLevel, msg)
 			}
 		}
 	}