--- a/log.go
+++ b/log.go
@@ -70,8 +70,8 @@
 //     // Create the hook:
 //     type SeverityHook struct{}
 //
-//     func (h SeverityHook) Run(e *zerolog.Event, level zerolog.Level, hasLevel bool, msg string) {
-//          if hasLevel {
+//     func (h SeverityHook) Run(e *zerolog.Event, level zerolog.Level, msg string) {
+//          if level != zerolog.NoLevel {
 //              e.Str("severity", level.String())
 //          }
 //     }
@@ -110,6 +110,8 @@ const (
 	FatalLevel
 	// PanicLevel defines panic log level.
 	PanicLevel
+	// NoLevel defines an absent log level.
+	NoLevel
 	// Disabled disables the logger.
 	Disabled
 )
@@ -128,6 +130,8 @@ func (l Level) String() string {
 		return "fatal"
 	case PanicLevel:
 		return "panic"
+	case NoLevel:
+		return ""
 	}
 	return ""
 }
@@ -233,28 +237,28 @@ func (l Logger) Hook(h Hook) Logger {
 //
 // You must call Msg on the returned event in order to send the event.
 func (l *Logger) Debug() *Event {
-	return l.newEvent(DebugLevel, true, nil)
+	return l.newEvent(DebugLevel, nil)
 }
 
 // Info starts a new message with info level.
 //
 // You must call Msg on the returned event in order to send the event.
 func (l *Logger) Info() *Event {
-	return l.newEvent(InfoLevel, true, nil)
+	return l.newEvent(InfoLevel, nil)
 }
 
 // Warn starts a new message with warn level.
 //
 // You must call Msg on the returned event in order to send the event.
 func (l *Logger) Warn() *Event {
-	return l.newEvent(WarnLevel, true, nil)
+	return l.newEvent(WarnLevel, nil)
 }
 
 // Error starts a new message with error level.
 //
 // You must call Msg on the returned event in order to send the event.
 func (l *Logger) Error() *Event {
-	return l.newEvent(ErrorLevel, true, nil)
+	return l.newEvent(ErrorLevel, nil)
 }
 
 // Fatal starts a new message with fatal level. The os.Exit(1) function
@@ -262,7 +266,7 @@ func (l *Logger) Error() *Event {
 //
 // You must call Msg on the returned event in order to send the event.
 func (l *Logger) Fatal() *Event {
-	return l.newEvent(FatalLevel, true, func(msg string) { os.Exit(1) })
+	return l.newEvent(FatalLevel, func(msg string) { os.Exit(1) })
 }
 
 // Panic starts a new message with panic level. The message is also sent
@@ -270,7 +274,7 @@ func (l *Logger) Fatal() *Event {
 //
 // You must call Msg on the returned event in order to send the event.
 func (l *Logger) Panic() *Event {
-	return l.newEvent(PanicLevel, true, func(msg string) { panic(msg) })
+	return l.newEvent(PanicLevel, func(msg string) { panic(msg) })
 }
 
 // WithLevel starts a new message with level.
@@ -290,6 +294,8 @@ func (l *Logger) WithLevel(level Level) *Event {
 		return l.Fatal()
 	case PanicLevel:
 		return l.Panic()
+	case NoLevel:
+		return l.Log()
 	case Disabled:
 		return nil
 	default:
@@ -302,9 +308,7 @@ func (l *Logger) WithLevel(level Level) *Event {
 //
 // You must call Msg on the returned event in order to send the event.
 func (l *Logger) Log() *Event {
-	// We use panic level with addLevelField=false to make Log passthrough all
-	// levels except Disabled.
-	return l.newEvent(PanicLevel, false, nil)
+	return l.newEvent(NoLevel, nil)
 }
 
 // Print sends a log event using debug level and no extra field.
@@ -335,23 +339,18 @@ func (l Logger) Write(p []byte) (n int, err error) {
 	return
 }
 
-func (l *Logger) newEvent(level Level, addLevelField bool, done func(string)) *Event {
+func (l *Logger) newEvent(level Level, done func(string)) *Event {
 	enabled := l.should(level)
 	if !enabled {
 		return nil
 	}
-	lvl := InfoLevel
-	if addLevelField {
-		lvl = level
-	}
-	e := newEvent(l.w, lvl, true)
+	e := newEvent(l.w, level, true)
 	e.done = done
-	e.hasLevel = addLevelField
 	if l.context != nil && len(l.context) > 0 && l.context[0] > 0 {
 		// first byte of context is ts flag
 		e.buf = json.AppendTime(json.AppendKey(e.buf, TimestampFieldName), TimestampFunc(), TimeFieldFormat)
 	}
-	if addLevelField {
+	if level != NoLevel {
 		e.Str(LevelFieldName, level.String())
 	}
 	if l.context != nil && len(l.context) > 1 {