Hi @rs  I improved my implementation (#24), making it faster and removing all allocations.

Since this adds a change to the Event struct (adds the hasLevel field), I'm sending a separate PR.
Please check (and hopefully merge) #24 before this one.

Thanks!