@rcoelho 2017-11-27
Some benchmarks:

Before the improvements:

```
BenchmarkHooks/Nop/Single-8             20000000                56.3 ns/op            40 B/op          2 allocs/op
BenchmarkHooks/Nop/Multi-8              20000000                81.8 ns/op            80 B/op          3 allocs/op
BenchmarkHooks/Simple-8                 20000000                91.3 ns/op            40 B/op          2 allocs/op
```

After the first commit:

```
BenchmarkHooks/Nop/Single-8             30000000                35.4 ns/op            16 B/op          1 allocs/op
BenchmarkHooks/Nop/Multi-8              30000000                45.1 ns/op            32 B/op          1 allocs/op
BenchmarkHooks/Simple-8                 30000000                56.6 ns/op            16 B/op          1 allocs/op
```

Aftere the second commit:

```
BenchmarkHooks/Nop/Single-8             50000000                22.4 ns/op             0 B/op          0 allocs/op
BenchmarkHooks/Nop/Multi-8              50000000                25.8 ns/op             0 B/op          0 allocs/op
BenchmarkHooks/Simple-8                 30000000                42.5 ns/op             0 B/op          0 allocs/op
```