@rs 2017-11-27
Do you think we could leverage the `Event.done` property which is already executed in `Msg` and `Msgf` to implement this?