@rs 2017-11-27
Will you address other comments in a subsequent commit?