@rcoelho 2017-11-28
> Do you think we could leverage the Event.done property which is already executed in Msg and Msgf to implement this?

Possible, but would complicate things. I think that it is good the way it is now.
Ready to merge if all changes look good to you.